Split out classes and use OverlayContainer.
