+++ osu.Framework/osu.Framework.csproj
@@ -135,6 +135,8 @@
     <Compile Include="Graphics\Transformations\TransformScaleVector.cs" />
     <Compile Include="Graphics\UserInterface\Button.cs" />
     <Compile Include="Graphics\UserInterface\TextBox.cs" />
+    <Compile Include="Graphics\Visualisation\FlashyBox.cs" />
+    <Compile Include="Graphics\Visualisation\InfoOverlay.cs" />
     <Compile Include="Graphics\Visualisation\TreeContainer.cs" />
     <Compile Include="Graphics\Visualisation\VisualisedDrawable.cs" />
     <Compile Include="Input\IKeyboardState.cs" />
